[front] Now registration adds new user at last. TODO: handle redirect to user's home page after successful registration, handle login server side validation and register server side validation - now I can send multiple POSTs to backend with same data and there are duplicate users :(
